rework of the rpc-tests with zombienet and individual js-scripts
