--- rpc-tests/astar-dev.toml
@@ -1,69 +1,0 @@
-[settings]
-timeout = 1000
-
-[relaychain]
-default_command = "./bin/polkadot"
-default_args = [ "-l=parachain=debug,xcm=trace" ]
-chain = "rococo-local"
-
-  [[relaychain.nodes]]
-  name = "relaychain-alice"
-  validator = true
-
-  [[relaychain.nodes]]
-  name = "relaychain-bob"
-  validator = true
-
-  [[relaychain.nodes]]
-  name = "relaychain-charlie"
-  validator = true
-
-  [[relaychain.nodes]]
-  name = "relaychain-dave"
-  validator = true
-
-[[parachains]]
-id = 2006
-chain = "astar-dev"
-cumulus_based = true
-
-  [[parachains.collators]]
-  name = "col-ast-ali"
-  command = "./bin/astar-collator"
-  ws_port = 9988
-  rpc_port = 8545
-  args = [ ]
-
-  [[parachains.collators]]
-  name = "col-ast-bob"
-  command = "./bin/astar-collator"
-  args = [ ]
-
-[[parachains]]
-id = 2007
-chain = "shiden-dev"
-cumulus_based = true
-
-  [[parachains.collators]]
-  name = "col-sdn-ali"
-  command = "./bin/astar-collator"
-  ws_port = 9989
-  rpc_port = 8546
-  args = [ ]
-
-  [[parachains.collators]]
-  name = "col-sdn-bob"
-  command = "./bin/astar-collator"
-  args = [ ]
-
-[[hrmp_channels]]
-  sender = 2006
-  recipient = 2007
-  max_capacity = 8
-  max_message_size = 512
-
-[[hrmp_channels]]
-  sender = 2007
-  recipient = 2006
-  max_capacity = 8
-  max_message_size = 512
