+++ rpc-tests/rpc-tests.toml
@@ -1,0 +1,52 @@
+[settings]
+timeout = 1000
+
+[relaychain]
+default_command = "./polkadot"
+chain = "rococo-local"
+
+  [[relaychain.nodes]]
+  name = "relay01"
+
+  [[relaychain.nodes]]
+  name = "relay02"
+
+  [[relaychain.nodes]]
+  name = "relay03"
+
+  [[relaychain.nodes]]
+  name = "relay04"
+
+[[parachains]]
+id = 2006
+chain = "astar-dev"
+cumulus_based = true
+
+  [parachains.collator]
+  name = "astar"
+  command = "./astar-collator"
+  rpc_port = 8545
+  args = [ "-l=xcm=trace", "--enable-evm-rpc" ]
+
+[[parachains]]
+id = 2007
+chain = "shiden-dev"
+cumulus_based = true
+
+  [parachains.collator]
+  name = "shiden"
+  command = "./astar-collator"
+  rpc_port = 8546
+  args = [ "-l=xcm=trace", "--enable-evm-rpc" ]
+
+[[hrmp_channels]]
+  sender = 2006
+  recipient = 2007
+  max_capacity = 8
+  max_message_size = 512
+
+[[hrmp_channels]]
+  sender = 2007
+  recipient = 2006
+  max_capacity = 8
+  max_message_size = 512
