+++ rpc-tests/rpc-tests.zndsl
@@ -1,0 +1,27 @@
+Description: RPC Tests
+Network: ./rpc-tests.toml
+Creds: config
+
+{% set nodes = ["relay01", "relay02", "relay03", "relay04"] %}
+{% for node in nodes %}
+  {{node}}: is up
+  {{node}}: reports node_roles is 4
+{% endfor %}
+
+{% set collators = ["shiden", "astar"] %}
+{% for collator in collators %}
+  # Basic Tests
+  {{collator}}: js-script ./tests/0001-fetch-chain-from-rpc-node.js return is equal to 1 within 200 seconds
+  {{collator}}: js-script ./tests/0001-fetch-name-from-rpc-node.js return is equal to 1 within 200 seconds
+  {{collator}}: js-script ./tests/0001-transfer-tokens.js with "Alice,Dave" return is equal to 1 within 200 seconds
+  {{collator}}: js-script ./tests/0001-regsiter-contract-on-h160.js with "Alice" within 200 seconds
+
+  # EVM Test Section
+  {{collator}}: js-script ./tests/0001-transfer-tokens-to-evm-mapped-account.js with "Alice" within 200 seconds
+  {{collator}}: js-script ./tests/0002-deploy-an-evm-contract return is equal to 1 within 200 seconds
+  {{collator}}: js-script ./tests/0002-get-evm-contract-data return is equal to 1 within 200 seconds
+  {{collator}}: js-script ./tests/0002-set-evm-contract-data with "Hola" return is equal to 1 within 200 seconds
+
+  # Shell script
+  {{collator}}: run ./tests/0001-custom.sh within 200 seconds
+{% endfor %}
