+++ rpc-tests/tests/0001-custom.sh
@@ -1,0 +1,2 @@
+# just echo for this test
+echo "OK"
